requirements.txt
moviepy
gtts
requests
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pytest
pillow
ffmpeg-python
pydantic
groq
